// src/utils/gmx/context/SyntheticsEvents/types.ts
export type SetPendingOrder = (data: PendingOrderData) => void;
// ...
export type PendingOrderData = {
  account: string;
  marketAddress: string;
  initialCollateralTokenAddress: string;
  swapPath: string[];
  initialCollateralDeltaAmount: BigNumber;
  minOutputAmount: BigNumber;
  sizeDeltaUsd: BigNumber;
  isLong: boolean;
  shouldUnwrapNativeToken: boolean;
  orderType: OrderType;
};